# Add type parameter for CachingOptimizer