--- a/perf/cachingoptimizer.jl
+++ b/perf/cachingoptimizer.jl
@@ -1,4 +1,6 @@
-# See https://github.com/JuliaOpt/MathOptInterface.jl/issues/321 and https://github.com/JuliaOpt/MathOptInterface.jl/pull/323
+# See https://github.com/JuliaOpt/MathOptInterface.jl/issues/321,
+#     https://github.com/JuliaOpt/MathOptInterface.jl/pull/323 and
+#     https://github.com/JuliaOpt/MathOptInterface.jl/pull/390
 
 using BenchmarkTools
 using MathOptInterface
@@ -10,7 +12,7 @@ optimizer = MOIU.MockOptimizer(Model{Float64}())
 caching_optimizer = MOIU.CachingOptimizer(Model{Float64}(), optimizer)
 MOIU.resetoptimizer!(caching_optimizer) # detach optimizer
 v = MOI.addvariables!(caching_optimizer, 2)
-cf = MOI.ScalarAffineFunction(v, [0.0, 0.0], 0.0)
+cf = MOI.ScalarAffineFunction(MOI.ScalarAffineTerm.([0.0, 0.0], v), 0.0)
 c = MOI.addconstraint!(caching_optimizer, cf, MOI.Interval(-Inf, 1.0))
 @btime MOI.modifyconstraint!($caching_optimizer, $c, $(MOI.Interval(0.0, 2.0)))
 MOIU.attachoptimizer!(caching_optimizer)