This PR adds a type parameter for the optimizer of `CachingOptimizer`. It adds this type parameters in the first place as it will be more useful than the second one (see e.g. `dropoptimizer!`). It also changes the order of the cache and the optimizer fields for consistency.
Note that it does not implement the idea of https://github.com/JuliaOpt/MathOptInterface.jl/issues/321#issuecomment-382729573 of removing the possibility to have no optimizer or to change the optimizer.
Indeed, this would require JuMP to handle these cases and I initially thought that this was the solution because I didn't have the idea of simply using the non concrete type parameter `Union{Nothing, MOI.AbstractOptimizer}` in the case where we want to be able to drop the optimizer and to change it and use the concrete optimizer type when we don't want to change it and the type instability becomes the bottleneck (e.g. https://github.com/JuliaOpt/MathOptInterface.jl/issues/321).

We can now make a similar change in JuMP. Have a parameter for the `moibackend` type that the user can set it to the concrete type of the optimizer to have no overhead (but loosing the features of changing optimizer, droping it, ...)

## Benchmarks
See https://github.com/JuliaOpt/MathOptInterface.jl/pull/323 for benchmarks before this change

With a CachingOptimizer with concrete field type for the optimizer field (i.e. `MockOptimizer{Model{Float64}}`):
```
v0.6
 36.591 ns (4 allocations: 112 bytes)
148.012 ns (8 allocations: 224 bytes)
 24.322 ns (3 allocations: 96 bytes)
 24.871 ns (3 allocations: 96 bytes)
v0.7
 12.130 ns (1 allocation: 48 bytes)
103.417 ns (3 allocations: 112 bytes)
 13.545 ns (1 allocation: 48 bytes)
 11.785 ns
```
With a CachingOptimizer with `Union{Nothing, OptimizerType}` with concrete `OptimizerType` for the `optimizer field:
```
v0.6
 40.083 ns (4 allocations: 112 bytes)
177.939 ns (11 allocations: 288 bytes)
 29.122 ns (3 allocations: 96 bytes)
 25.401 ns (3 allocations: 96 bytes)
v0.7
 13.049 ns (1 allocation: 48 bytes)
109.670 ns (3 allocations: 112 bytes)
 11.501 ns (1 allocation: 48 bytes)
 12.421 ns (1 allocation: 48 bytes)
```

The overhead of the union with `Void` is negligible in v0.7 as suggested by https://github.com/JuliaOpt/MathOptInterface.jl/pull/390#discussion_r195424172 so it seems cleaner to use it

Closes https://github.com/JuliaOpt/MathOptInterface.jl/issues/321